Split javascript files in the iframe package (bug 1188357)
