+++ .gitignore
@@ -58,3 +58,4 @@ src/tests/marketplace-core-modules/
 tags
 tests/captures/*
 yulelog_*
+package/iframe/bundle.js
