+++ package.json
@@ -28,8 +28,10 @@
     "underscore": "1.7.0"
   },
   "devDependencies": {
+    "browserify": "^11.0.1",
     "casperjs": "^1.1.0-beta3",
     "chai": "^2.0.0",
+    "envify": "^3.4.0",
     "karma": "^0.12.31",
     "karma-chai": "^0.1.0",
     "karma-firefox-launcher": "^0.1.4",
@@ -41,6 +43,7 @@
     "phantomjs": "1.9.7-15",
     "requirejs": "^2.1.16",
     "sinon": "^1.12.2",
-    "slimerjs": "^0.9.5"
+    "slimerjs": "^0.9.5",
+    "vinyl-source-stream": "^1.1.0"
   }
 }
