+++ package/iframe/index.html
@@ -27,6 +27,7 @@
         <button class="global button try-again" data-l10n="try-again">Try again</button>
       </div>
     </div>
-  <script type="text/javascript" src="main.js"></script>
+    <iframe id="iframe" seamless></iframe>
+    <script src="bundle.js"></script>
   </body>
 </html>
